maths/requirements.txt
pymupdf==1.27.1
pillow==12.1.1
